Incomplete correction of MemoItem onDrag logic.
Previous version did not account for the case when the dragOffset covered multiple items in the list e.g a drag from 6 to 2 would only update the positions of 2 and 6 but not for the items between.
The previous behavior would appear correct when the dragOffset was in steps for example from 2 to 3, 3 to 4, 4 to 5, 5 to 6. However, it would only appear to be correct and the items would never perfectly align with the desired position.
Correct behavior is only partly implemented as processingOverDrag still does not account for the case where the item's initial index is between start and end indices.